Move redux stuff into component
