+++ src/App.js
@@ -1,8 +1,5 @@
-import { createSlice, configureStore, createSelector } from "@reduxjs/toolkit";
-import { useEffect } from "react";
-import { counterSlice } from "./Store";
+import ReduxDemonstration from "./ReduxDemonstration";
 import "./App.css";
-import { useDispatch, useSelector } from "react-redux";
 
 // ROUTING
 const routes = [
@@ -21,10 +18,7 @@ const renderNavButtons = () => {
 
 // BASE APP
 function App() {
-    const dispatch = useDispatch();
-  const { add, minus } = counterSlice.actions;
 
-  const reduxValue = useSelector(state => state.value)
 
   return (
     <div className="app">
@@ -34,24 +28,7 @@ function App() {
         </div>
       </section>
       <section className="body">
-        <div>
-          <label for="add">Add 1</label>
-          <input
-            type="button"
-            id="add"
-            name="add"
-            label="+"
-            onClick={() => dispatch(add())}
-          />
-          <label for="add">Subtract 1</label>
-          <input
-            type="button"
-            id="minus"
-            name="minus"
-            onClick={() => dispatch(minus())}
-          />
-          <div>Value: {reduxValue}</div>
-        </div>
+        <ReduxDemonstration />
       </section>
       <section className="footer">footer</section>
     </div>
